mt_eval/requirements.txt
# Galen Translation Evaluation Pipeline Requirements
# Installation Instructions:
# 1. Create virtual environment: python -m venv venv
# 2. Activate it:
#    - Windows: venv\Scripts\activate
#    - Mac/Linux: source venv/bin/activate
# 3. Install dependencies: pip install -r requirements.txt

# ============================================
# Translation APIs
# ============================================
openai>=1.0.0              # OpenAI GPT models
anthropic>=0.7.0           # Claude models
google-genai>=0.2.0        # Gemini models

# ============================================
# Core Utilities
# ============================================
python-dotenv>=1.0.0       # Environment variable management
requests>=2.28.0           # HTTP requests

# ============================================
# Evaluation Metrics - Lexical
# ============================================
nltk>=3.8                  # BLEU-4, METEOR (downloads: punkt, wordnet, omw-1.4)
sacrebleu>=2.3.0           # chrF++ (character + word n-grams)
rouge-score>=0.1.2         # ROUGE-L

# ============================================
# Evaluation Metrics - Neural/Semantic
# ============================================
bert-score>=0.3.13         # BERTScore (contextual embeddings)
unbabel-comet>=2.0.0       # COMET (crosslingual evaluation)

# BLEURT - must be installed separately from GitHub:
# pip install git+https://github.com/google-research/bleurt.git
# Note: COMET and BLEURT require downloading large models (~1-2GB total)

# ============================================
# Data Processing
# ============================================
numpy>=1.24.0
pandas>=2.0.0

# ============================================
# Optional: GPU acceleration
# ============================================
# torch>=2.0.0             # PyTorch with CUDA for faster neural metrics
# If using GPU, install torch separately with CUDA support:
# pip install torch --index-url https://download.pytorch.org/whl/cu118
